Trait luminance::framebuffer::ColorSlot

pub trait ColorSlot<C, L, D> where C: HasFramebuffer + HasTexture, L: Layerable, D: Dimensionable, D::Size: Copy {
    fn color_formats() -> Vec<PixelFormat>;
    fn reify_textures(size: D::Size, mipmaps: usize, textures: &mut Vec<C::ATexture>) -> Self;
}

A framebuffer has a color slot. A color slot can either be empty (the unit type is used,()) or several color formats.

Required Methods

fn color_formats() -> Vec<PixelFormat>

Turn a color slot into a list of pixel formats.

fn reify_textures(size: D::Size, mipmaps: usize, textures: &mut Vec<C::ATexture>) -> Self

Reify a list of raw textures into a color slot.