Module luminance::pipeline

Dynamic rendering pipelines.

This module gives you materials to build dynamic rendering pipelines. A Pipeline represents a functional stream that consumes geometric data and rasterizes them.

Structs

Pipeline
A dynamic rendering pipeline. A pipeline is responsible of rendering into a Framebuffer.

RenderCommand
A render command, which holds information on how to rasterize tessellation.

ShadingCommand
A dynamic shading command. A shading command gathers render commands under a shader Program.

Traits

HasPipeline
Trait to implement to add Pipeline support.

SomeShadingCommand
This trait is used to add existential quantification to ShadingCommands. It should be implemented by backends to enable their use in Pipelines.