Struct luminance::pipeline::Pipeline

pub struct Pipeline<'a, C, L, D, CS, DS> where C: 'a + HasFramebuffer + HasProgram + HasTessellation + HasTexture, L: 'a + Layerable, D: 'a + Dimensionable, D::Size: Copy, CS: 'a + ColorSlot<C, L, D>, DS: 'a + DepthSlot<C, L, D> {
    pub framebuffer: &'a Framebuffer<C, L, D, CS, DS>,
    pub clear_color: [f32; 4],
    pub shading_commands: Vec<&'a SomeShadingCommand>,
}

A dynamic rendering pipeline. A pipeline is responsible of rendering into a Framebuffer.

L refers to the Layering of the underlying Framebuffer.

D refers to the Dim of the underlying Framebuffer.

CS and DS are – respectively – the color and depth Slot of the underlying Framebuffer.

Fields

framebuffer

clear_color

shading_commands

Methods

impl<'a, C, L, D, CS, DS> Pipeline<'a, C, L, D, CS, DS> where C: HasPipeline, L: Layerable, D: Dimensionable, D::Size: Copy, CS: ColorSlot<C, L, D>, DS: DepthSlot<C, L, D>

fn new(framebuffer: &'a Framebuffer<C, L, D, CS, DS>, clear_color: [f32; 4], shading_commands: Vec<&'a SomeShadingCommand>) -> Self

fn run(&self)

Run a Pipeline.