Trait luminance::shader::program::HasProgram

pub trait HasProgram: HasStage + HasUniform {
    type Program;
    fn new_program(tess: Option<(&Self::AStage, &Self::AStage)>, vertex: &Self::AStage, geometry: Option<&Self::AStage>, fragment: &Self::AStage) -> Result<Self::Program, ProgramError>;
    fn free_program(program: &mut Self::Program);
    fn map_uniform(program: &Self::Program, name: String) -> Result<Self::U, ProgramError>;
    fn update_uniforms<F>(program: &Self::Program, f: F) where F: Fn();
}

Trait to implement to provide shader program features.

Associated Types

type Program

Required Methods

fn new_program(tess: Option<(&Self::AStage, &Self::AStage)>, vertex: &Self::AStage, geometry: Option<&Self::AStage>, fragment: &Self::AStage) -> Result<Self::Program, ProgramError>

Create a new program by linking it with stages.

fn free_program(program: &mut Self::Program)

Free a program.

fn map_uniform(program: &Self::Program, name: String) -> Result<Self::U, ProgramError>

Map a uniform name to its uniform representation. Can fail with ProgramError.

fn update_uniforms<F>(program: &Self::Program, f: F) where F: Fn()

Bulk update of uniforms. The update closure should contain the code used to update as many uniforms as wished.