Enum luminance::shader::stage::StageError

pub enum StageError {
    CompilationFailed(Type, String),
    UnsupportedType(Type),
}

Variants
CompilationFailed(Type, String)

Occurs when a shader fails to compile.

UnsupportedType(Type)

Occurs when you try to create a shader which type is not supported on the current hardware.