Module luminance::shader::uniform

Shader uniforms and associated operations.

Uniforms kick in several and useful ways. They’re used to customize shaders.

Structs

Uniform
A shader uniform. Uniform<C, T> doesn’t hold any value. It’s more like a mapping between the host code and the shader the uniform was retrieved from.

UniformUpdate
Wrapper over Uniform, discarding everything but update.

Enums

Dim
Dimension of the uniform.

Type
Type of a uniform.

Traits

HasUniform

Uniformable
Layering of a uniform (flat or layered/array). Types that can behave as Uniform.