Trait luminance::tessellation::HasTessellation

pub trait HasTessellation {
    type Tessellation;
    fn new<T>(mode: Mode, vertices: &[T], indices: Option<&[u32]>) -> Self::Tessellation where T: Vertex;
    fn destroy(tessellation: &mut Self::Tessellation);
}

Trait to implement to provide tessellation features.

Associated Types

type Tessellation

A type representing tessellation on GPU.

Required Methods

fn new<T>(mode: Mode, vertices: &[T], indices: Option<&[u32]>) -> Self::Tessellation where T: Vertex

Create a Tessellation from its vertices and a Mode.

If indices == None, the vertices represent an array of vertices that are connected to each others in the order they appear. If you want to connect them in another way, you can index them with Some(indices).

fn destroy(tessellation: &mut Self::Tessellation)

Destroy a Tessellation.