Struct luminance::texture::Sampler

pub struct Sampler {
    pub wrap_r: Wrap,
    pub wrap_s: Wrap,
    pub wrap_t: Wrap,
    pub minification: Filter,
    pub magnification: Filter,
    pub depth_comparison: Option<DepthComparison>,
}

A Sampler object gives hint on how a Texture should be sampled.

Fields

wrap_r

How should we wrap around the r sampling coordinate?

wrap_s

How should we wrap around the s sampling coordinate?

wrap_t

How should we wrap around the t sampling coordinate?

minification

Minification filter.

magnification

Magnification filter.

depth_comparison

For depth textures, should we perform depth comparison and if so, how?