Trait luminance::vertex::Vertex

pub trait Vertex {
    fn vertex_format() -> VertexFormat;
}

A type that can be used as a Vertex has to implement that trait – it must provide a mapping to VertexFormat.

If you’re not sure on how to implement that or if you want to use automatic types, feel free to use the primary supported types and Chain or tuples.

Required Methods

fn vertex_format() -> VertexFormat