Module luminance::texture

This module provides texture features.

Structs

Cubemap

Dim1

Dim2

Dim3

Flat

Layered

Sampler
A Sampler object gives hint on how a Texture should be sampled.

Texture
Texture.

Enums

CubeFace
Faces of a cubemap.

DepthComparison
Depth comparison to perform while depth test. a is the incoming fragment’s depth and b is the fragment’s depth that is already stored.

Dim
Dimension of a texture.

Filter
Minification and magnification filter.

Layering
Texture layering. If a texture is layered, it has an extra coordinates to access the layer.

Wrap
How to wrap texture coordinates while sampling textures?

Traits

Dimensionable
Reify a type into a Dim.

HasTexture
Trait to implement to provide texture features.

Layerable
Reify a type into a Layering.

Functions

dim_capacity